+++ realgui/widget/gui_glass/gui_glass.c
@@ -696,7 +696,7 @@ uint16_t gui_glass_get_width(gui_glass_t *_this)
     else if (_this->storage_type == IMG_SRC_FTL)
     {
         struct gui_rgb_data_head head;
-        uint32_t base = (uint32_t)(uintptr_t)_this->data;
+        uint32_t base = (uintptr_t)_this->data;
         gui_ftl_read(base, (uint8_t *)&head, sizeof(gui_rgb_data_head_t));
         return head.w;
     }
@@ -729,7 +729,7 @@ uint16_t gui_glass_get_height(gui_glass_t *_this)
     else if (_this->storage_type == IMG_SRC_FTL)
     {
         struct gui_rgb_data_head head;
-        uint32_t base = (uint32_t)(uintptr_t)_this->data;
+        uint32_t base = (uintptr_t)_this->data;
         gui_ftl_read(base, (uint8_t *)&head, sizeof(gui_rgb_data_head_t));
         return head.h;
     }
